pyproject.toml
[project]
name = "avochoc-assessment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.128.0",
    "passlib>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
